Files deletion

+++ backend/src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -55,7 +59,10 @@ export class UsersService {
    * @param id User unique id
    * @returns Request to remove user from database
    */
-  async remove(id: string): Promise<UserDocument> {
+  async remove(id: string, reqId: string): Promise<UserDocument> {
+    const user = await this.userModel.findById(reqId);
+    if (!user.isAdmin || user._id != id)
+      throw new UnauthorizedException('You dont have permission to do that!');
     return this.userModel.findByIdAndDelete(id).exec();
   }
 
@@ -71,6 +78,16 @@ export class UsersService {
       throw new ForbiddenException(
         'Sorry, you have reached your storage limit. You have used up all the available space in your account. Free up some space by deleting files or asking administrator for more space.',
       );
+    return this.userModel
+      .findByIdAndUpdate(userId, user)
+      .setOptions({ overwrite: true, new: true });
+  }
+
+  async removalOfFile(userId, fileSize) {
+    const user = await this.userModel.findById(userId);
+
+    user.storedData -= fileSize;
+    if (user.storedData < 0) user.storedData = 0;
     return this.userModel
       .findByIdAndUpdate(userId, user)
       .setOptions({ overwrite: true, new: true });
